+++ examples/sveltekit-openai/src/routes/api/completion/+server.ts
@@ -1,5 +1,5 @@
 import { createOpenAI } from '@ai-sdk/openai';
-import { StreamData, StreamingTextResponse, streamText } from 'ai';
+import { StreamData, streamText } from 'ai';
 
 import { env } from '$env/dynamic/private';
 // You may want to replace the above with a static private env variable
@@ -17,24 +17,19 @@ export const POST = (async ({ request }) => {
   // Extract the `prompt` from the body of the request
   const { prompt } = await request.json();
 
+  // optional: use stream data
+  const data = new StreamData();
+  data.append({ test: 'value' });
+
   // Ask OpenAI for a streaming chat completion given the prompt
   const result = await streamText({
     model: openai('gpt-4-turbo-preview'),
     prompt,
-  });
-
-  // optional: use stream data
-  const data = new StreamData();
-
-  data.append({ test: 'value' });
-
-  // Convert the response into a friendly text-stream
-  const stream = result.toAIStream({
-    onFinal(completion) {
+    onFinish() {
       data.close();
     },
   });
 
   // Respond with the stream
-  return new StreamingTextResponse(stream, {}, data);
+  return result.toAIStreamResponse({ data });
 }) satisfies RequestHandler;