+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/google-vertex
 
+## 0.0.16
+
+### Patch Changes
+
+- 0eabc798: feat (provider/google-vertex): change vertexai library into peer dependency
+
 ## 0.0.15
 
 ### Patch Changes